pyproject.toml
[project]
name = "drf-blog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-taggit>=6.1.0",
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "djangorestframework-simplejwt>=5.4.0",
    "pillow>=11.1.0",
    "django-anymail>=12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]
prod = [
    "gunicorn>=23.0.0",
]
